geFixtureForm refactor
